+++ src/metrics_test.go
@@ -48,9 +48,48 @@ func createGoldenFile(i *integration.Integration, sourceFile string) (string, []
 	actualContents, _ := i.Entities[0].Metrics[0].MarshalJSON()
 
 	if *update {
-		ioutil.WriteFile(goldenFile, actualContents, 0644)
+		_ = ioutil.WriteFile(goldenFile, actualContents, 0644)
 	}
 	return goldenFile, actualContents
+}
+
+func TestGetLocalNodeID(t *testing.T) {
+	fakeClient := mockClient{}
+	mockedReturnVal := filepath.Join("testdata", "good-nodes-local.json")
+	fakeClient.On("Request", localNodeInventoryEndpoint).Return(mockedReturnVal, nil).Once()
+
+	nodeID, err := getLocalNodeID(&fakeClient)
+	assert.NoError(t, err)
+	assert.Equal(t, "z9ZPp87vT92qG1cRVRIcMQ", nodeID)
+}
+
+func TestGetLocalNodeID_Error(t *testing.T) {
+	fakeClient := mockClient{}
+	mockedReturnVal := filepath.Join("testdata", "bad-nodes-local.json")
+	fakeClient.On("Request", localNodeInventoryEndpoint).Return(mockedReturnVal, nil).Once()
+
+	_, err := getLocalNodeID(&fakeClient)
+	assert.Error(t, err)
+	assert.Equal(t, errLocalNodeID, err)
+}
+
+func TestGetMasterNodeID(t *testing.T) {
+	fakeClient := mockClient{}
+	mockedReturnVal := filepath.Join("testdata", "good-master.json")
+	fakeClient.On("Request", electedMasterNodeEndpoint).Return(mockedReturnVal, nil).Once()
+
+	nodeID, err := getMasterNodeID(&fakeClient)
+	assert.NoError(t, err)
+	assert.Equal(t, "z9ZPp87vT92qG1cRVRIcMQ", nodeID)
+}
+
+func TestGetMasterNodeID_Error(t *testing.T) {
+	fakeClient := mockClient{}
+	fakeClient.On("Request", electedMasterNodeEndpoint).Return("", nil).Once()
+
+	_, err := getMasterNodeID(&fakeClient)
+	assert.Error(t, err)
+	assert.Equal(t, errMasterNodeID, err)
 }
 
 func TestPopulateNodesMetrics(t *testing.T) {
@@ -58,7 +97,8 @@ func TestPopulateNodesMetrics(t *testing.T) {
 	client := createNewTestClient()
 	client.init("nodeStatsMetricsResult.json", nodeStatsEndpoint, t)
 
-	populateNodesMetrics(i, client, testClusterName)
+	err := populateNodesMetrics(i, client, testClusterName)
+	assert.NoError(t, err)
 
 	sourceFile := filepath.Join("testdata", "nodeStatsMetricsResult.json")
 	goldenFile, actualContents := createGoldenFile(i, sourceFile)
@@ -87,7 +127,9 @@ func TestPopulateClusterMetrics(t *testing.T) {
 	client := createNewTestClient()
 	client.init("clusterStatsMetricsResult.json", clusterEndpoint, t)
 
-	populateClusterMetrics(i, client, "")
+	name, err := populateClusterMetrics(i, client, "")
+	assert.NotEmpty(t, name)
+	assert.NoError(t, err)
 
 	sourceFile := filepath.Join("testdata", "clusterStatsMetricsResult.json")
 
@@ -121,7 +163,9 @@ func TestPopulateCommonMetrics(t *testing.T) {
 	args.CollectPrimaries = true
 	client.init("commonMetricsResult.json", commonStatsEndpoint, t)
 
-	populateCommonMetrics(i, client, testClusterName)
+	name, err := populateCommonMetrics(i, client, testClusterName)
+	assert.NotEmpty(t, name)
+	assert.NoError(t, err)
 
 	sourceFile := filepath.Join("testdata", "commonMetricsResult.json")
 	goldenFile, actualContents := createGoldenFile(i, sourceFile)
@@ -154,9 +198,11 @@ func TestPopulateIndicesMetrics(t *testing.T) {
 
 	commonStruct := new(CommonMetrics)
 	commonData, _ := ioutil.ReadFile(filepath.Join("testdata", "indicesMetricsResult_Common.json"))
-	json.Unmarshal(commonData, commonStruct)
+	err := json.Unmarshal(commonData, commonStruct)
+	assert.NoError(t, err)
 
-	populateIndicesMetrics(i, client, commonStruct, testClusterName)
+	err = populateIndicesMetrics(i, client, commonStruct, testClusterName)
+	assert.NoError(t, err)
 
 	sourceFile := filepath.Join("testdata", "indicesMetricsResult.json")
 	goldenFile, actualContents := createGoldenFile(i, sourceFile)
@@ -206,9 +252,11 @@ func TestIndicesRegex(t *testing.T) {
 
 	commonStruct := new(CommonMetrics)
 	commonData, _ := ioutil.ReadFile(filepath.Join("testdata", "indicesMetricsResult_Common.json"))
-	json.Unmarshal(commonData, commonStruct)
+	err := json.Unmarshal(commonData, commonStruct)
+	assert.NoError(t, err)
 
-	populateIndicesMetrics(i, client, commonStruct, testClusterName)
+	err = populateIndicesMetrics(i, client, commonStruct, testClusterName)
+	assert.NoError(t, err)
 
 	actualLength := len(i.Entities)
 	expectedLength := 1