+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.3.5 (2021-04-20)
+### Added
+- option `master_only` affection `command: all` only: Collect cluster metrics on the elected master only
+- Upgraded github.com/newrelic/infra-integrations-sdk to v3.6.7
+- Switched to go modules
+- Upgraded pipeline to go 1.16
+- Replaced gometalinter with golangci-lint
+
 ## 4.3.4 (2021-03-23)
 ### Changed
 - Add arm packages and binaries.